Milestone: Design is tested enough in simulation to try it out on the real hardware!
altera jtag_uart_0.vhd to be able to read data from the scope while the debugger is on!

Added: Netlist simulation
Improoved: solved the most hold time violations in the netlist simulation

connected the RX of the generic uart to the external trigger input.
altera jtag_uart_0.vhd to be able to read data from the scope while the debugger is on!

added DSO_FrontPanel.c for leds and buttons functional tests after the sram problem is solved!

Bugfixes: 
SpecialFunctionRegister-ea.vhd wrong data could be read
corrected the baudrate in Leon3Uart.c
corrected and prooved the pin assignments


git-svn-id: https://welecw2000a.svn.sourceforge.net/svnroot/welecw2000a@84 5642ed1f-f172-45da-a169-216d43dc357c